Move Cards from Styled Folder
